+++ src/bos/server/controllers/v2/sessiontemplates.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2021-2024 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2021-2025 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -23,6 +23,7 @@
 #
 import logging
 import connexion
+from connexion.lifecycle import ConnexionResponse
 
 from bos.common.utils import exc_type_msg
 from bos.server.models.v2_session_template import V2SessionTemplate as SessionTemplate  # noqa: E501
@@ -138,8 +139,8 @@ def get_v2_sessiontemplate(session_template_id):
     if session_template_id not in DB:
         LOGGER.warning("Session template not found: %s", session_template_id)
         return connexion.problem(
-            status=404, title="Sessiontemplate could not found.",
-            detail="Sessiontemplate {} could not be found".format(session_template_id))
+            status=404, title="Session template not found.",
+            detail="Session template {} could not be found".format(session_template_id))
     template = DB.get(session_template_id)
     return template, 200
 
@@ -166,8 +167,8 @@ def delete_v2_sessiontemplate(session_template_id):
     if session_template_id not in DB:
         LOGGER.warning("Session template not found: %s", session_template_id)
         return connexion.problem(
-            status=404, title="Sessiontemplate could not found.",
-            detail="Sessiontemplate {} could not be found".format(session_template_id))
+            status=404, title="Session template not found.",
+            detail="Session template {} could not be found".format(session_template_id))
     return DB.delete(session_template_id), 204
 
 
@@ -183,8 +184,8 @@ def patch_v2_sessiontemplate(session_template_id):
     if session_template_id not in DB:
         LOGGER.warning("Session template not found: %s", session_template_id)
         return connexion.problem(
-            status=404, title="Sessiontemplate could not found.",
-            detail="Sessiontemplate {} could not be found".format(session_template_id))
+            status=404, title="Session template not found.",
+            detail="Session template {} could not be found".format(session_template_id))
 
     if connexion.request.is_json:
         LOGGER.debug("connexion.request.is_json")
@@ -231,10 +232,13 @@ def validate_v2_sessiontemplate(session_template_id: str):
     a session from being launched using this template.
     """
     LOGGER.debug("GET /v2/sessiontemplatesvalid/%s invoked validate_v2_sessiontemplate", session_template_id)
-    data, status_code = get_v2_sessiontemplate(session_template_id)
+    response = get_v2_sessiontemplate(session_template_id)
+    if isinstance(response, ConnexionResponse):
+        # This means it was an error, so we just pass it up
+        return response
 
-    if status_code != 200:
-        return data, status_code
+    # Otherwise it should be a tuple of data and 200 status code
+    data, _ = response
 
     # We assume boot because it and reboot are the most demanding from a validation
     # standpoint.